Added test cor notes: db model, db controller, and server controller.
